flip read2 sense/anti for locus plotter plotting
Locus plotter parses the first column and always assigns the data from the row containing "sense" on the top so for visualizing the endo cut site, we need to flip the labels. I've added awk commands to do so for each bulk processing script marking read2 in the tag pileups.